{"schema":"decodex/commit/1","summary":"Move frozen capture handoff runtime","authority":"manual"}
